add remove_when_older variable to check_create_new_file (#81)

+++ hhnk_research_tools/general_functions.py
@@ -64,21 +64,38 @@ def convert_gdb_to_gpkg(gdb, gpkg, overwrite=False, verbose=True):
 def check_create_new_file(
     output_file: Union[Path, str],
     overwrite: bool = False,
-    input_files: list = [],
+    input_files: Union[list[Path], None] = None,
+    remove_when_older: bool = True,
     check_is_file=True,
 ) -> bool:
     """
     Check if we should continue to create a new file.
 
-    output_file: The file under evaluation
-    overwrite: When overwriting is True the output_file will be removed.
-    input_files: if input files are provided the edit time of the input will be
-                 compared to the output. If edit time is after the output, it will
-                 recreate the output.
-    check_is_file: check if output_file is a file
+    Parameters
+    ----------
+    output_file : Path | str
+        The file under evaluation
+    overwrite : bool, default is True
+        When overwriting is True the output_file will be removed.
+    input_files : list[Path], default is None
+        When input files are provided the edit time of the input will be
+        compared to the output. If edit time is after the output, it will
+        recreate the output.
+    remove_when_older : bool, default is True
+        There may be scenario's where this automatic removal of
+        older output_files is unwanted. Changing this to False,
+        will skip that step. Make sure the file is deleted / overwritten
+        when create
+    check_is_file bool, default is True
+        When True, raises TypeError if output_file is not a file
+
+    Returns
+    -------
+    create : bool,
+        Bool if the inputfile should be created.
     """
     create = False
-    output_file = Path(str(output_file))  # noqa
+    output_file = Path(str(output_file))
 
     # Raise when no suffix if checking for file
     if check_is_file:
@@ -102,10 +119,15 @@ def check_create_new_file(
                 input_mtime = input_file.stat().st_mtime
 
                 if input_mtime > output_mtime:
-                    logger.info(
-                        f"check_create_new_file - Removed {output_file.name} due to creation time older than input"
-                    )
-                    output_file.unlink()
+                    if remove_when_older:
+                        logger.info(
+                            f"check_create_new_file - Removed {output_file.name} due to creation time older than input"
+                        )
+                        output_file.unlink()
+                    else:
+                        logger.info(
+                            f"check_create_new_file - File {output_file.name} has creation time older than input. But was not removed."
+                        )
                     create = True
                     break
     return create
